Move ClientProfile enum into constants
