+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/resources/META-INF/resources/admin/init.jsp
@@ -39,7 +39,7 @@ page import="com.liferay.oauth2.provider.service.OAuth2AuthorizationServiceUtil"
 page import="com.liferay.oauth2.provider.web.internal.constants.OAuth2AdminWebKeys" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AssignScopesModel"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.AuthorizationModel"%><%@
-page import="com.liferay.oauth2.provider.web.internal.display.context.ClientProfile"%><%@
+page import="com.liferay.oauth2.provider.web.internal.constants.ClientProfile"%><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2AdminPortletDisplayContext" %><%@
 page import="com.liferay.oauth2.provider.web.internal.display.context.OAuth2ApplicationsManagementToolbarDisplayContext" %><%@
 page import="com.liferay.oauth2.provider.web.internal.portlet.OAuth2AdminPortlet"%><%@
